[refac] move package oauth to external/openfeign
