+++ src/main/java/org/hankki/hankkiserver/api/auth/service/AuthService.java
@@ -11,14 +11,14 @@ import org.hankki.hankkiserver.domain.user.model.Platform;
 import org.hankki.hankkiserver.common.exception.EntityNotFoundException;
 import org.hankki.hankkiserver.common.exception.InvalidValueException;
 import org.hankki.hankkiserver.common.exception.UnauthorizedException;
-import org.hankki.hankkiserver.oauth.apple.AppleOAuthProvider;
-import org.hankki.hankkiserver.oauth.dto.SocialInfoDto;
-import org.hankki.hankkiserver.oauth.kakao.KakaoOAuthProvider;
+import org.hankki.hankkiserver.external.openfeign.apple.AppleOAuthProvider;
+import org.hankki.hankkiserver.external.openfeign.dto.SocialInfoDto;
 import org.hankki.hankkiserver.domain.user.repository.UserInfoRepository;
 import org.hankki.hankkiserver.domain.user.repository.UserRepository;
 import org.hankki.hankkiserver.api.auth.controller.request.UserLoginRequest;
 import org.hankki.hankkiserver.api.auth.service.response.UserLoginResponse;
 import org.hankki.hankkiserver.api.auth.service.response.UserReissueResponse;
+import org.hankki.hankkiserver.external.openfeign.kakao.KakaoOAuthProvider;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
