+++ src/main/java/org/hankki/hankkiserver/external/openfeign/apple/AppleIdentityTokenParser.java
@@ -1,4 +1,4 @@
-package org.hankki.hankkiserver.oauth.apple;
+package org.hankki.hankkiserver.external.openfeign.apple;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
