+++ src/main/java/org/hankki/hankkiserver/external/openfeign/apple/dto/ApplePublicKey.java
@@ -1,4 +1,4 @@
-package org.hankki.hankkiserver.oauth.apple.dto;
+package org.hankki.hankkiserver.external.openfeign.apple.dto;
 
 import java.util.Optional;
