+++ src/main/java/org/hankki/hankkiserver/external/openfeign/apple/dto/AppleRevokeRequest.java
@@ -1,4 +1,4 @@
-package org.hankki.hankkiserver.oauth.apple.dto;
+package org.hankki.hankkiserver.external.openfeign.apple.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
