+++ src/main/java/org/hankki/hankkiserver/external/openfeign/kakao/dto/KakaoUserInfo.java
@@ -1,10 +1,11 @@
-package org.hankki.hankkiserver.oauth.kakao.dto;
+package org.hankki.hankkiserver.external.openfeign.kakao.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserProfile(
-        String nickname
+public record KakaoUserInfo(
+        Long id,
+        KakaoAccount kakaoAccount
 ) {
 }
