+++ src/main/java/org/hankki/hankkiserver/external/openfeign/kakao/dto/KakaoUserProfile.java
@@ -1,11 +1,10 @@
-package org.hankki.hankkiserver.oauth.kakao.dto;
+package org.hankki.hankkiserver.external.openfeign.kakao.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public record KakaoUserInfo(
-        Long id,
-        KakaoAccount kakaoAccount
+public record KakaoUserProfile(
+        String nickname
 ) {
 }
